INTERLOK-292 Refactored the interfaces. INTERLOK-292 #time 2h
Changed the pluginFactory behaviour so that it returns a ClientPlugin
which contains the receivers. Modified the aspects so that it
uses getPlugin().getReceivers() when firing events.

Modified the DateFormat so that it's more ISO8601 by default (it was
coming out with the actual 'Z' character as well, so it was somewhat
wrong anyways.