INTERLOK-292 Refactoring for maintainability. INTERLOK-292 #time 3h
Refactoring code for less code duplication. All the aspects now extend
a BaseAspect which provides some common code.
Refactored ReflectionHelper to try and get an ID from both getUniqueId
and getId; getUniqueId is preferred. That's probably a throwback to v2
days when not everything had a getUniqueId; still we're doing it via
reflection so most of the time it won't trigger the 2nd method.

Moved StepIncrementor into the aspect package as it's only used there.